Add Log Viewer to MainUI (#2884)
Closes #2883.
This adds a basic log viewer into the Main UI, and displays the logs
from openHAB log file using a websocket connection.

## Overview 

This adds the log viewer to the Developer Tools -:

The main logger window is reasonably simple, and a little busy. Log
entries are styled with a background color:
red=error, yellow=warn, greed=info, grey text = trace, no change = debug.
The window scrolls horizontally - the time column
remains fixed, but the log level, class and message scroll left/right.
This allows the user to effectively remove these columns if they desire
by simply setting the horizontal scroll point accordingly.

Across the top in the navbar and the bottom toolbar are a bunch of status and configuration
data/buttons. These are:

* Play button: Continues receiving/processing logs if paused / stopped.
* Pause: Continues to receive logs from openHAB, but doesn't add them to
the display, so the display is static. When Play is pressed, the filter
is updated, so any recent entries will be added.
* Stop: Stops receiving data from the server. The websocket connection
is disconnected - this ensures there is no change to the local buffer.
* Filter input box - this filters the display to display only text in
either the logger name or message containing the filter text (case
insensitive).
* Time of the oldest and latest log entries held in the local buffer.
* Number of log entries currently displayed and in the buffer.
The color of the badge is green if all are displayed, orange if some are
filtered out, or red if the max size of the buffer (2000) is reached.
* Download: Downloads a CSV of the filtered log file for further
processing.
* Copy: Copies the log as HTML to the clipboard. This allows pasting of
the filtered log complete with highlighting.
* Delete: Clears the log buffer.
* Exclamation: Shows `ERROR` logs even if they would otherwise be
filtered out.
* Pen: Configures highlighting (more on this below).
* Cog: Configured logging (more on this below).
 
## Highlighting 

The UI allows the user to add any number of text elements that will be
highlighted. The user can specify a colour, and enable and disable if
the highlighting is active or not. The list of highlights is saved
locally.

## Log Settings 

The user can directly manipulate log levels. This allows the user to
simply change the log level for existing packages, or delete the log
entry (restoring it to default) or adding a new logger for a custom
package. If adding a custom package, this gets added as `WARN` level (it
uses the log level provided in the `ROOT` logger), which can then be
changed in the dropdown.

Also-by: Yannick Schaus <github@schaus.net>
Signed-off-by: Chris Jackson <chris@cd-jackson.com>