Log viewer: only render lines around visible area
This should fix the performance problems in the log viewer when the number of items reaches the max buffer size.

It works by calculating the range of line indexes currently visible in the scroll area and redraw only those along with those around them (50 on each side). The rest are replaced by 2 padded empty spaces of the appropriate heights, so that the table appears to be of the expected total height.

Signed-off-by: Yannick Schaus <github@schaus.net>