Model: Don't subscribe to Item events to reload
This avoids that the global SSE connection is always established,
and therefore reduces problems with the max parallel HTTP connections limitation.
The global SSE connection is now only established if the web audio sink or the command Item are enabled.

Whilst this change potentially makes the UI not reload the model automatically on change,
practically the now removed mechanism might haven't been really useful at all because the model is already reloaded on Main UI start,
and if SSE was not connected the moment an Item changed the mechanism did not trigger.

Signed-off-by: Florian Hotze <dev@florianhotze.com>