Tag Input: Fix automatic add on blur (#2982)
User inadvertently losing the entered new tag when they haven't realised
that they needed to press Enter. It seems that onblur was originally
intended to auto-add the typed tag, e.g. when user immediately clicked
"Save" (top right), but it was attached to the wrong element, thus the
event didn't fire.

Changes keypress to keyup event because the keypress is
missing the last letter otherwise. It wasn't a problem in the past
because it only worked with Enter key (being the last keypress). Missing
the last letter becomes a problem when adding it in the onblur event. So
keyup solved that.

Also removed `@keyPressed.native="keyPressed"`, not sure why it was needed.

Signed-off-by: Jimmy Tanagra <jcode@tanagra.id.au>