Update slider and stepper handing of step
Fixes #3019, see that bug for precise details on what led to this
change.

Updating oh-slider and oh-stepper to round the state of the linked Item
based on the step provided (from the StateDescription). Before this
change, a step value such as 1.234 would try to fix the value like:

Number(...).toFixed(234)

which causes a RangeError (toFixed requires inputs randing 0..100).

This change attempts to keep as close to the current behavior, and the
intention mentioned by the comments. In the above example it fix the
state to 3 digits (instead of 234).

Testing:

Confirmed that this addresses the console errors listed in the bug as
well as makes the slider and steppers interactive again when using very
precise step values.

Tested with integer steps (fixing to 0 decmal digits), undefined steps
(fixing to 0 decimal digits), and very precise steps generated from my
mqtt.generic items (0.3937007874015748, fixing to 16).

The slider will default to toFixed(0) if no step is provided, but the
stepper will continue to skip fixing the state entirely when a step
isn't provided. Ideally it seems that these should behave the same, but
scoping the change to limit impact.

Signed-off-by: David Tschida <dmtschida1@gmail.com>