Retry loading of rules-list
During startup the rule engine can take some time to start. If rules-list is opened too early, the following message will appear: "The rule engine must be installed before rules can be created.". This commit makes sure that loading is retried until it succeeds, so that the message disappears when the rule engine is ready.

Signed-off-by: Ravi Nadahar <nadahar@rediffmail.com>