Sitemap editor: fix negative number parameters (#3055)
See: https://community.openhab.org/t/basic-ui-setpoint-negative-values/161011

Parameters with negative number where not parsed correctly. This is
because hyphen is defined as a token and therefore already recognized
before recognizing the number. Changing the order does not help as it
breaks chart period parsing.

This fix avoids recognizing numbers as tokens and handles numbers in the
lexer.

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>