Add rule-edit "Source" tab
Rules that are generated by external scripting add-ons and that use SimpleRule or other ScriptModuleTypeProvider based jsr223 implementations "cheats" in that they don't actually provide the content of the module, but instead embeds a "hidden reference" to a runnable object in memory that is executed when the module is called. These in-memory runnable objects exist as bytecode and can't be displayed by OH in any meaningful way. To be able to view what the rule does from the UI, the "source" of the rule in the scripting language in question can be embedded in the rule as metadata. For rules with such embedded "source" metadata, a new tab is created that shows this rule source.

Signed-off-by: Ravi Nadahar <nadahar@rediffmail.com>