Add Alt-Shift-M shortcut key to toggle menu when unpinned (#3081)
It has always been a pain point when the menu is not pinned,
getting back to it is very difficult, involving several "back" clicks,
and sometimes it also gets lost and it became impossible without a full
browser refresh.

This adds a shortcut Alt-Shift-M (similar to Alt-Shift-D) to show the
menu whereever you are.

Signed-off-by: Jimmy Tanagra <jcode@tanagra.id.au>