Replace GH Action from tj-actions by custom code
The GitHub Actions build workflow includes external Actions from 3rd party repositories.
The functionality of tj-actions/verify-changed-files was used to detect modifications of files under source control during the build process.
The external action is removed and replaced by custom code.
This is less flexible, but does not require an external action in the build toolchain.

Currently, tj-actions is not available and anyway no longer allowed by our GH Actions configuration.
This PR needs to be picked to every active branch (or branches to be rebased).

Signed-off-by: Holger Friedrich <mail@holger-friedrich.de>