Make Main UI log levels configurable (#3226)
As console logging consumes a considerable amount of time, even when
the browser console is closed (especially since logging objects requires
deep cloning them), implemented a UI wide log level.
The log level allows to only log to the console when the level is above a
set value. These levels mimic the OH server side log levels.

The log level for the browser can be set in an extra developer tools
screen entry, and are persisted in the browser store when set.
The default logging level is INFO (will include ERROR and WARN), thereby
disabling `console.log`, `console.debug` and `console.trace`.

The impact is UI wide.

---------

Also-by: Florian Hotze <dev@florianhotze.com>
Signed-off-by: Mark Herwege <mark.herwege@telenet.be>