Model treeview: Fixes for drag and drop (#3223)
This fixes 2 issues:

- When picking items from model, the already checked or selected items
would not appear if these are non-semantic. It would also return a bad
selection (selecting the already selected semantic item twice). This PR
fixes it by enabling showNonSemantic if any of the selected or checked
items are non-semantic and expanding the tree to show all selected or
checked items.
- Drag/drop sometimes fails when moving into a subtree of a sibling of a
node being moved, as the added and removed events where not generated,
but a moved event is created instead. This fix also interprets the moved
event. To make things more easily understood for the users when
dragging, the ghost element is now hidden, as it sometimes showed in
another place than where it would be dropped.
- Drag/drop performance on large models was sometimes very slugish. Some
performance improvements where made that should improve this.

---------

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>