Location card: Various improvements and fixes (#3245)
This does:

1. The collection methods can now not only include exact Property /
Equipment values but also (_**new**_) children of those values
(optionally). This allows display of not only Lightbulb and Lightstripe,
but also any LightSource or child thereof. And also of doors and windows
that support either OpenState (digital) or OpenLevel (analog) positions;
to name a couple.
2. Aligns the code in this PR with the
[documentation](https://github.com/openhab/openhab-docs/pull/2514) and
vice versa; taking into account
[#556](https://github.com/openhab/openhab-webui/issues/556) and
subsequent PRs related to that. This PR and #3239 added support for
Motion detectors under presence, other displays under screens.
3. Cleans up the collection of lighting points to prevent duplicate
entries in the location card; that may otherwise arise when (say) the
user connects a Switch, a Dimmer, and a Color item to the same light
channel.

Resolves
https://github.com/openhab/openhab-webui/pull/3239#issuecomment-3004118138

---------

Also-by: Mark Herwege <mark.herwege@telenet.be>
Signed-off-by: Andrew Fiddian-Green <software@whitebear.ch>