Use versioned branch for add-on & sidebar docs (#3254)
Closes #2945.

Load the markdown source files used by the add-on details and the
developer sidebar from the versioned git branch instead of always using
the latest docs. This is implemented by loading the docs src from the
website, were Netlify redirects are used to point each version to the
proper branch.

---------

Signed-off-by: Florian Hotze <dev@florianhotze.com>