Add page to edit semantic tags tree (#3165)
This adds a page with a treeview of the existing semantic tags in the
development tools and allows adding (and editing) custom semantic tags.
With this, you can maintain custom tags from the UI (it was already
possible from `tags.yaml` files before). Moving around custom tags with
drag and drop is possible.

---------

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>