Add to model from Thing: Allow picking existing items and groups (#3200)
Resolves #433.

Adding to the model from thing channels has been enhanced in a few ways:

- Add Points to Model: now allows to pick an existing item already
linked to the channel. The item will get default semantic tags from the
channel defaults if it does not have semantic tags already.
- Add Equipment to Model: ability to pick an existing group for the
equipment. This group will get the default equipment semantic tag from
the thing if it does not have a semantic tag already. Groups tagged as
locations cannot be selected to guard model consistency. Adding points
in this case is enhanced just as above.

When doing this, I ran into some reactivity issues with oh-icon, where
switching icon and state at the same time caused update issues. This was
caused by conflicts in the 2 watchers updating the same properties. I
fixed that by grouping these 2 elements in 1 object with a single
watcher.

---------

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>