Add treeview semantics picker (#3178)
Closes https://github.com/openhab/openhab-webui/issues/3211.

This PR changes the semantics picker to use a treeview.

This treeview is inherited from
https://github.com/openhab/openhab-webui/pull/3165 and enhanced.
It therefore also introduces some enhancements in the semantic treeview maintenance.

The semantics treeview presents a searchable treeview of the semantic
classes/properties, limited to what is relevant in the context. When
used as a picker, radio buttons are used for picking. It has filter and
expand capabilities, options to show the semantic name and synonyms and
tooltips showing the semantic descriptions if available.


---------

Also-by: Florian Hotze <dev@florianhotze.com>
Signed-off-by: Mark Herwege <mark.herwege@telenet.be>