Fix ECharts animation regression (#3321)
#3136 led to a strange issue with ECharts animations.
When values are updated, the animations don't animate from the previous value,
but from "zero" (or whatever the initial value is).
This breaks some widgets, like it was reported in https://community.openhab.org/t/barometer-gauge-widget-3-1-0-0-6-0-0-0/161270.

Ultimately, this has been likely traced back to some implementation detail in vue-echarts,
see the PR for more details.

Signed-off-by: Ravi Nadahar <nadahar@rediffmail.com>