Copy File Definition: Work around the security error in Safari (#3309)
Reported in
https://community.openhab.org/t/copy-dsl-yaml-definition-not-working-for-me/165233

Tests the clipboard copy functionality across browsers.
If the operation fails, e.g. in Safari due to platform restrictions, an
additional dialog is shown after the async fetch completes. This ensures
that when the user clicks the "OK" button, the clipboard action executes
immediately using the pre-fetched data, complying with Safari's
transient activation requirements.

The additional dialog is only shown in browsers that have such
restriction.

---------

Signed-off-by: Jimmy Tanagra <jcode@tanagra.id.au>