Sitemap editor: Fix sitemap save (#3339)
The fix provided with https://github.com/openhab/openhab-webui/pull/3300
was incomplete. It only removed the parent field from the first level of
the tree.

This fix iterates over the levels.
It also removes the JSON parse/stringify combination as it was used to
copy the object and is replaced by a deepClone.

This should be backported to 5.0 as the backported fix provided did not
work in all cases.

---------

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>