[blockly] change event attributes, add return block (#3376)
Depends on: https://github.com/openhab/openhab-addons/pull/19443

The event object/attributes in script actions and conditions is now the
same as in JSRule and Rule Builder.
Backwards compatibility will be retained for some time, though it is
advised to adjust, which is done in this PR according to:

https://github.com/openhab/openhab-js/blob/51f164dc9f11a29e6223e24268848aa8f0be168f/src/rules/rules.js#L471

Additionally **a new return block** was added to the "Run & Process" category with return parameter.

1) The block without a return parameter is a nice improvement to easily
leave a script without using non-elegant if/else statements. Bei adding
the return statement the script is immediately ended. In this case the
return parameter is not important.
2) By adding a return parameter (in particular true or false) this can be used for script conditions

---------

Signed-off-by: Stefan Höhn <mail@stefanhoehn.com>