[blockly] fix zdt copy_of nano handling (#3361)
Fixes #3188 

The root of the problem is that the underlying time library does not
support to set the nanoseconds only. Nanoseconds also contain the millis
and the micros. As a result, for example changing only microseconds
requires to merge the new microseconds between the millis and the actual
nano seconds. The has now been fixed in the copy-of method that allows
to overwrite them individually. 

As a side effect or natural consequence, _nanoTotal_ was introduced to
retrieve the >whole< nanoseconds including millis an micros, whereas
_nano_ only retrieves the actual nanoseconds.

**Important side note or warning:** 

When retrieving the current datetime with _now_ there is a precision
limitation within javascript that will only return a time with precision
milliseconds! This means that micros and nanos are always set to zero.
This is not an openHAB limitation but one that relates to javascript
only.

Signed-off-by: Stefan Höhn <mail@stefanhoehn.com>