Pin icons: Make "pinned" state icon solid and "unpinned" state outlined. (#3288)
The way the "pin" icons are currently displayed:

When an item is pinned, the pin icon is shown as red with slash. This
makes me wonder: is this item currently pinned or not?

So instead of the icon showing what will happen when it's clicked, show
the current state (is it currently pinned or unpinned). Clicking it will
toggle it.

Furthermore, when it's unpinned, show it as gray color, and when it's
pinned, show it as blue, indicating that "yes it's pinned".

-----

Signed-off-by: Jimmy Tanagra <jcode@tanagra.id.au>