oh-slider: Protect against posting errant slider updates (#3444)
Some of my sliders would occasionally be sending a NaN value back to
openHAB when the position was updating.
I'm not sure if this was present before Vue upgrade in #3350, but i have noticed we now pop up an error
dialog anytime a command receives a non `2xx` response, so its a lot more apparent if there are issues.

Signed-off-by: Dan Cunningham <dan@digitaldan.com>