Main UI: Visualize semantic model tree in item details screen (#3227)
Closes https://github.com/openhab/openhab-webui/issues/3217

This PR extends the item details screen to show part of the model the
item is in, from the top, down to all its children. The tree is opened
up to the level of the item. Clicking on a level in the tree, will take
you to the related item.

With this PR, the terminology discussion from
https://github.com/openhab/openhab-webui/issues/3211 becomes less
critical as the terminology is replaced by a graphical representation of
the model with all information in it.

---------

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>