Model: Show non semantic children in model tree (#3476)
Closes https://github.com/openhab/openhab-webui/issues/3330

Non semantic items and groups will be made visible as children of
semantic groups in the model tree when shown non-semantic is switched
on. Semantic children will not be shown to avoid having semantic items
visible multiple times in the model tree.

Some improvements where made to drag and drop:
- solved issue with tree not being updated after moving (not sure this
resolves all issues, but it does resolve what I could reproduce)
- drag & drop failed when something was selected at the start of
dragging: fixed by deselecting at start of drag and fix highlighting of
selected node, make selection work as toggle on the model page
- scroll dragged node into view after dragging

---------

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>