States store: Performance improvements for pages with many Items (#3501)
Fixes #1746.

When loading pages with many (e.g. >= 100 Items), opening the page takes
a lot of time.
This is because for every Item in that page, the `trackedItems` proxy
adds that Item to the tracking list, causing a ton of reactivity to be
triggered (e.g. widget rerendering, expression evaluation).

To fix this, batch processing for wanted tracking list Items is
introduced.

When a new page is loaded now, there will still be many Item states
missing. But instead of directly adding them to the tracking list, now
they are added to the pending new Items Set. This Set is then processed
by a interval at most 100 ms later.

Further optimizations include not publishing internal state (such as
EventSource) and using reactivity only where needed.

---------

Also-by: Jeff James <jeff@james-online.com>
Signed-off-by: Florian Hotze <dev@florianhotze.com>