Charts: Fix duplicate REST requests (#3544)
Reported on the community:
https://community.openhab.org/t/5-1-0-m3-charts-load-slowly/167402/6

Fixes an issue where the chart-mixin did the same requests multiple
times due to a reactivity issues.
This fix makes the caching for items non-reactive by using
underscore-prefixed properties, avoiding recomputation of
`asyncComputed` `series` due to changes to the items store.
This fix also caches ongoing promises to `/rest/items/{itemname}` and
`/rest/persistence/items/{itemname}` to reduce the number of requests.

---------

Signed-off-by: Florian Hotze <dev@florianhotze.com>