Refactor config sheet advanced toggle (#3503)
Closes https://github.com/openhab/openhab-webui/issues/3480

As described in the issue, it is sometimes confusing advanced parameters
with a non-default value are shown also when the show advanced toggle is
off.

This refactors that behaviour:
- make the 'Show advanced' checkbox clearly toggle between showing the
advanced parameters or not (and hide the option if there are no advanced
options).
- put a badge next to the 'Show advanced' toggle that shows the number
of advanced parameters with a non-default value (not shown if all have
default values).
- the badges have tooltips clarifying their meaning.

---------

Signed-off-by: Mark Herwege <mark.herwege@telenet.be>