Analyzer/Charts: Fix week starting on Sunday starts on Monday (#3578)
The `week` chart type should range from Sunday to Sunday, but it has
been rendered similarly to `isoWeek` from Monday to Monday.

This fixes that behaviour by properly calculating the startTime and
removing the alignment logic from oh-time-axis, as startTime and endTime
are already properly aligned by the chart-mixin.

It also makes week starting on Sunday available in the analyzer.

---------

Also-by: Jeff James <jeff@james-online.com>
Signed-off-by: Florian Hotze <dev@florianhotze.com>