oh-plan-page: Make marker tooltips non interactive (#3612)
On mobile devices the plan markers are very hard to interact with 
as the label pops up on touch events, and prevents the plan marker 
from reacting to the event. 
This means you have to try and touch around it, or zoom 
in and touch which is not ideal. 
This change makes the labels non interactive (they still appear on touch or hover).

Signed-off-by: Dan Cunningham <dan@digitaldan.com>