`oh-*-card` widgets: Fix style applied to both oh-card and child component (#3677)
- `oh-*-card`: The `style` config should only be applied to the
parent/root element, but no children.
- Widget components: Do not v-bind `$attrs` to child element(s):
`$attrs` automatically binds to the root element of the component, so do
not v-bind to children.
This for example applies style to the components and all its children,
which isn't correct.
- `oh-block`: Do not v-bind `$attrs` to child element(s) & Remove
wrapping div as f7-block already is a div.

---------

Signed-off-by: Florian Hotze <dev@florianhotze.com>
(cherry picked from commit 43e7adb2bbafd79776838237058096df6c87034d)